refactor(qbittorrent): move torrent state into client module
